lab4 refactoring
+division into files-components
+some component reworks
+ {} in if
